requirements.txt
flask>=3.0
requests>=2.31
pywebview>=5.0
